+++ lib/versioncake/view_additions.rb
@@ -3,11 +3,11 @@ require 'action_view'
 # register an addition detail for the lookup context to understand,
 # this will allow us to have the versions available upon lookup in
 # the resolver.
-ActionView::LookupContext.register_detail(:versions){ VersionCake::Railtie.config.versioncake.supported_versions }
+ActionView::LookupContext.register_detail(:versions){ [] }
 
 ActionView::PathResolver.class_eval do
   # not sure why we are doing this yet, but looks like a good idea
-  if ActionPack::VERSION::MAJOR >= 4 && ActionPack::VERSION::MINOR >= 1
+  if ActionPack::VERSION::MAJOR >= 4 && ActionPack::VERSION::MINOR >= 1 || ActionPack::VERSION::MAJOR >= 5
     ActionView::PathResolver::EXTENSIONS.replace({
                                                      locale: ".",
                                                      formats: ".",