+++ spec/rails_helper.rb
@@ -36,4 +36,6 @@ RSpec.configure do |config|
   # the seed, which is printed after each run.
   #     --seed 1234
   config.order = "random"
+
+  config.include VersionCake::TestHelpers, type: :controller
 end