+++ spec/unit/strategies/path_parameter_strategy_spec.rb
@@ -5,15 +5,13 @@ describe VersionCake::PathParameterStrategy do
   subject { strategy.extract(request) }
 
   context "a request with an api_version path parameter retrieves the version" do
-    let(:request) { instance_double('Request', path_parameters: \
-      {api_version: '11', other: 'parameter'}) }
+    let(:request) { instance_double('Request', path: 'api/v11/parameter') }
 
     it { is_expected.to eq 11 }
   end
 
   context "a request without an api_version path parameter returns nil" do
-    let(:request) { instance_double('Request', path_parameters: \
-      {other: 'parameter', another: 'parameter'}) }
+    let(:request) { instance_double('Request', path: 'parameter/parameter') }
 
     it { is_expected.to be_nil }
   end